Added/Updated: styles/general.css and index.html
Added styles directory to store the styling for the challenge.
Added general.css to set general styles for the challenge.
Updated index.html to link general.css to the file and updated the attribution section to include my name
and link to my GitHub profile and set the rel tag for the links.